+++ packages/refine/src/components/Dropdowns/WorkloadDropdown/index.tsx
@@ -1,25 +1,20 @@
 import { Icon, Menu } from '@cloudtower/eagle';
-import {
-  Pause16GradientBlueIcon,
-  RecoverContinue16GradientBlueIcon,
-} from '@cloudtower/icons-react';
+import { Retry16GradientBlueIcon } from '@cloudtower/icons-react';
 import { useResource, useUpdate, useCan } from '@refinedev/core';
-import { CronJob } from 'kubernetes-types/batch/v1';
 import React from 'react';
 import { useTranslation } from 'react-i18next';
 import { AccessControlAuth } from 'src/constants/auth';
-import { CronJobModel } from '../../models';
-import { pruneBeforeEdit } from '../../utils/k8s';
+import { WorkloadModel } from '../../../models';
+import { pruneBeforeEdit } from '../../../utils/k8s';
 import K8sDropdown, { DropdownSize } from '../K8sDropdown';
 
-type Props<Model extends CronJobModel> = {
+type Props<Model extends WorkloadModel> = {
   record: Model;
   size?: DropdownSize;
 };
 
-export function CronJobDropdown<Model extends CronJobModel>(props: Props<Model>) {
-  const { record, size } = props;
-  const { spec } = record as CronJob;
+export function WorkloadDropdown<Model extends WorkloadModel>(props: React.PropsWithChildren<Props<Model>>) {
+  const { record, size, children } = props;
   const { resource } = useResource();
   const { mutateAsync } = useUpdate();
   const { t } = useTranslation();
@@ -28,17 +23,14 @@ export function CronJobDropdown<Model extends CronJobModel>(props: Props<Model>)
     action: AccessControlAuth.Edit
   });
 
-  const suspended = Boolean(spec?.suspend);
-
   return (
     <K8sDropdown record={record} size={size}>
       {
         canEditData?.can !== false ? (
           <Menu.Item
             onClick={async () => {
-              const v = suspended ? record.resume() : record.suspend();
-              const id = record.id;
-
+              const v = record.redeploy();
+              const id = v.id;
               pruneBeforeEdit(v);
               await mutateAsync({
                 id,
@@ -46,21 +38,21 @@ export function CronJobDropdown<Model extends CronJobModel>(props: Props<Model>)
                 values: v,
                 successNotification() {
                   return {
-                    message: t(suspended ? 'dovetail.resume_success_toast' : 'dovetail.pause_success_toast', {
+                    message: t('dovetail.redeploy_success_toast', {
                       kind: record.kind,
-                      name: id,
+                      name: record.id,
                       interpolation: {
                         escapeValue: false,
                       }
                     }),
-                    type: 'success',
+                    type: 'success'
                   };
                 },
                 errorNotification() {
                   return {
-                    message: t(suspended ? 'dovetail.resume_failed_toast' : 'dovetail.pause_failed_toast', {
+                    message: t('dovetail.redeploy_failed_toast', {
                       kind: record.kind,
-                      name: id,
+                      name: record.id,
                       interpolation: {
                         escapeValue: false,
                       }
@@ -71,12 +63,12 @@ export function CronJobDropdown<Model extends CronJobModel>(props: Props<Model>)
               });
             }}
           >
-            <Icon src={suspended ? RecoverContinue16GradientBlueIcon : Pause16GradientBlueIcon}>
-              {t(suspended ? 'dovetail.resume' : 'dovetail.suspend')}
-            </Icon>
+            <Icon src={Retry16GradientBlueIcon}>{t('dovetail.redeploy')}</Icon>
           </Menu.Item>
+
         ) : null
       }
-    </K8sDropdown >
+      {children}
+    </K8sDropdown>
   );
 }
