+++ packages/refine/src/components/index.ts
@@ -6,15 +6,15 @@ export * from './ConditionsTable';
 export * from './FormLayout';
 export * from './FormErrorAlert';
 export * from './PodContainersTable';
-export * from './WorkloadDropdown';
-export * from './ReplicasDropdown';
-export * from './PodDropdown';
+export * from './Dropdowns/WorkloadDropdown';
+export * from './Dropdowns/ReplicasDropdown';
+export * from './Dropdowns/PodDropdown';
 export * from './CreateButton';
 export * from './ImageNames';
 export * from './ResourceCRUD';
 export * from './WorkloadPodsTable';
-export * from './CronJobDropdown';
-export * from './K8sDropdown';
+export * from './Dropdowns/CronJobDropdown';
+export * from './Dropdowns/K8sDropdown';
 export * from './DropdownMenuItems';
 export * from './ResourceUsageBar';
 export * from './WorkloadReplicas';
