+++ packages/refine/src/pages/cronjobs/list/index.tsx
@@ -1,7 +1,7 @@
 import { IResourceComponentsProps } from '@refinedev/core';
 import React from 'react';
 import { useTranslation } from 'react-i18next';
-import { CronJobDropdown } from 'src/components/CronJobDropdown';
+import { CronJobDropdown } from 'src/components/Dropdowns/CronJobDropdown';
 import { ListPage } from 'src/components/ListPage';
 import { Time } from 'src/components/Time';
 import { useEagleTable } from 'src/hooks/useEagleTable';
