+++ packages/refine/src/pages/deployments/list/index.tsx
@@ -2,7 +2,7 @@ import { IResourceComponentsProps } from '@refinedev/core';
 import React from 'react';
 import { useTranslation } from 'react-i18next';
 import { ListPage } from 'src/components/ListPage';
-import { ReplicasDropdown } from 'src/components/ReplicasDropdown';
+import { ReplicasDropdown } from 'src/components/Dropdowns/ReplicasDropdown';
 import { useEagleTable } from 'src/hooks/useEagleTable';
 import {
   AgeColumnRenderer,
